ARCHITECTURE CHANGES to summarizer.py
